Refactor custom actions validation logic
